Function ic_utils::fs::write_atomically

pub fn write_atomically<PDst, F>(dst: PDst, action: F) -> Result<()> where
    F: FnOnce(&mut BufWriter<&File>) -> Result<()>,
    PDst: AsRef<Path>, 

Atomically write to dst file, using a random file in the parent directory of dst as the temporary file.

Pre-conditions

dst is not a directory.

The parent directory of dst must be writeable.

Panics

Doesn't panic unless action panics.